+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ## Added
 
+- Letter Picker
 - Payee helper getPayeeIdFromTextMatchingOrAdd()
 
 ## Updated